// src/components/molecules/MemberItem/types.ts
export interface MemberItemParams {
  uid: string;
  isActive?: boolean;
  isSpeaker?: boolean;
  isModerator?: boolean;
  isRaising?: boolean;
  onClick: (id: string) => void;
  onClickMenu: (key: string) => void;
  speaker: string;
  profileUrl: string;
  emoji?: string;
  emojiId?: string;
  isLoggedIn?: boolean;
  isMuted?: boolean;
  micEnabled?: boolean;
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  peer?: any;
  role?: 'founding father' | 'founder' | 'pioneer' | 'basic';
}